+++ src/generators/mod_types/index.ts
@@ -1,5 +1,6 @@
 export { default as ElderMods } from './ElderMods';
 export { default as ShapedMods } from './ShapedMods';
 export { default as BestiaryAspectMods } from './BestiaryAspectMods';
+export { default as IncursionTempleMods } from './IncursionTempleMods';
 export { default as MasterSignatureMods } from './MasterSignatureMods';
 export { default as WarbandsMods } from './WarbandsMods';